attempt at adding modules, ran into a wall
This adds a ‘module <name>;’ syntax that then associates the module with any new roles/classes defined. The module is then compiled as if it is a namespace prefix. This works fine for class/role declarations being compiled into package declarations, but it doesn’t work so well for subclassing or role consumption. The modules store information about the classes/roles associated with it, and then the compiler uses this to decide if the superclass or role consumption need to have their namespaces adjusted because of a module in context. However, this module information is limited to a single compilation unit, which means that a module definition cannot span two files, which is not ideal :(

Gotta give it some more thought.